rp2/machine_rtc: add initial support for RTC.
Initial support for machine.RTC on rp2 port. It only supports datetime()
method and nothing else. The method gets/returns a tuple of 8 items,
just like esp32 port, for example, but the usec parameter is ignored as
the RP2 RTC only works up to seconds precision.

The Pico RTC isn't very useful as the time is lost during reset and
there seems to be no way to easily power up just the RTC clock with a
low current voltage, but still there seems to be use-cases for that, see
issues #6831, and a thonny issue #1592. It was also requested for
inclusion on v1.15 roadmap on #6832.

Signed-off-by: Krzysztof Adamski <k@japko.eu>